+++ go.mod
@@ -10,6 +10,7 @@ require (
 	github.com/newrelic/go-agent/v3 v3.9.0
 	github.com/newrelic/go-agent/v3/integrations/nrlambda v1.2.0
 	github.com/stretchr/testify v1.6.1
+	golang.org/x/mod v0.4.1
 	golang.org/x/net v0.0.0-20200904194848-62affa334b73 // indirect
 	golang.org/x/sys v0.0.0-20200909081042-eff7692f9009 // indirect
 	golang.org/x/text v0.3.3 // indirect
